+++ CHANGELOG.md
@@ -8,7 +8,9 @@
 * `DataFrameSchema`: defines a complete serializable dataframe schema  
 * `DataFrameExpr`: represents a DataFrame as a SQL expression  
 * `DataFrameMethod`: represents a DataFrame as a method call with support for both Narwhals and Native API  
-* `DataFrameTransformer`: represents a DataFrame as a series of chained transformations  
+* `DataFrameTransformer`: represents a DataFrame as a series of chained transformations
+* `DLTPipeline`: configuration options
+* `DLTViewDataSink`: View inside a DLT pipeline
 * `DTypes`: defines a serializable column data type
 * `UnityCatalogDataSource`: reads DataFrame from Unity Catalog 
 * `HiveMetastoreDataSource`: reads DataFrame from Hive Metastore